requirements/common.txt
Django==1.8.8
django-rest-swagger==0.3.5
djangorestframework==3.5.3
django-filter==1.0.1
#mysql-python==1.2.5
pycrypto==2.6.1
requests==2.18.4
xlrd==1.0.0
django-audit-log==0.7.0
Pillow==3.1.0
coreapi==2.3.0
django-cors-headers==2.1.0